+++ packages/snap/src/context.ts
@@ -152,8 +152,6 @@ const confirmationHandler = new ConfirmationHandler({
   snapClient,
   state,
   tronWebFactory,
-  assetsService,
-  feeCalculatorService,
 });
 
 /**